requirements.txt
absl-py                           1.2.0
albucore                          0.0.20
albumentations                    1.4.21
annotated-types                   0.7.0
astunparse                        1.6.3
backports.entry-points-selectable 1.1.1
cachetools                        5.5.0
certifi                           2021.10.8
charset-normalizer                2.0.12
contourpy                         1.3.1
cycler                            0.11.0
Cython                            0.29.30
descartes                         1.1.0
distlib                           0.3.4
eval_type_backport                0.2.0
filelock                          3.5.0
fire                              0.7.0
fonttools                         4.55.0
fsspec                            2024.10.0
gast                              0.5.3
google-pasta                      0.2.0
idna                              3.3
imageio                           2.31.6
imgaug                            0.4.0
Jinja2                            3.1.4
joblib                            1.4.2
Keras-Preprocessing               1.1.2
kiwisolver                        1.3.2
lazy_loader                       0.4
MarkupSafe                        3.0.2
matplotlib                        3.5.3
mpmath                            1.2.1
networkx                          3.4.2
numpy                             1.26.4
nuscenes-devkit                   1.1.11      /scratch/amtissot/nuscenes_data/nuscenes-devkit/python-sdk
nvidia-cublas-cu12                12.4.5.8
nvidia-cuda-cupti-cu12            12.4.127
nvidia-cuda-nvrtc-cu12            12.4.127
nvidia-cuda-runtime-cu12          12.4.127
nvidia-cudnn-cu12                 9.1.0.70
nvidia-cufft-cu12                 11.2.1.3
nvidia-curand-cu12                10.3.5.147
nvidia-cusolver-cu12              11.6.1.9
nvidia-cusparse-cu12              12.3.1.170
nvidia-nccl-cu12                  2.21.5
nvidia-nvjitlink-cu12             12.4.127
nvidia-nvtx-cu12                  12.4.127
opencv-python                     4.10.0.84
opencv-python-headless            4.10.0.84
opt-einsum                        3.3.0
packaging                         21.3
pandas                            1.4.2
parameterized                     0.9.0
Pillow                            10.0.1
pip                               22.0.4
platformdirs                      2.4.0
ply                               3.11
protobuf                          3.20.0
pycocotools                       2.0.8
pydantic                          2.10.2
pydantic_core                     2.27.1
pyparsing                         3.0.6
pyquaternion                      0.9.9
python-dateutil                   2.8.2
pytz                              2021.3
PyYAML                            6.0.2
requests                          2.26.0
scikit-image                      0.22.0
scikit-learn                      1.3.2
scipy                             1.13.1
semver                            2.8.1
setuptools                        58.1.0
Shapely                           1.8.5.post1
simsimd                           6.2.1
six                               1.16.0
sphire                            1.4.1
stringzilla                       3.11.0
sympy                             1.13.1
termcolor                         1.1.0
threadpoolctl                     3.5.0
tifffile                          2024.9.20
torch                             2.5.1
torchaudio                        2.5.1
torchvision                       0.20.1
tqdm                              4.64.1
triton                            3.1.0
ttach                             0.0.3
typing_extensions                 4.12.2
urllib3                           1.26.6
virtualenv                        20.10.0
wheel                             0.37.0
wrapt                             1.13.3
xarray                            2022.3.0